refactor(validation): group the list-param helpers
The folder is at its 12-item limit, and the object-path work below adds
another file to it. The three files about comma-separated list params —
lengths, pairing, targets — read as a set, so they move together and
leave room.
