+++ src/tools/actions/duplicate/helpers/clip/duplicate-clip-order-helpers.ts
@@ -7,8 +7,8 @@ import {
   isTakeLaneClip,
   type ArrangementTrack,
 } from "#src/tools/shared/arrangement/helpers/take-lane-helpers.ts";
-import { pairValues } from "#src/tools/shared/validation/list-pairing.ts";
-import { requireSameLength } from "#src/tools/shared/validation/list-lengths.ts";
+import { pairValues } from "#src/tools/shared/validation/lists/list-pairing.ts";
+import { requireSameLength } from "#src/tools/shared/validation/lists/list-lengths.ts";
 import { parseArrangementLength } from "../duplicate-helpers.ts";
 
 /**
