+++ src/tools/actions/duplicate/helpers/sources/duplicate-label-helpers.ts
@@ -15,7 +15,7 @@ import {
   getNameForIndex,
   parseNames,
 } from "#src/tools/shared/validation/name-utils.ts";
-import { type ListEntries } from "#src/tools/shared/validation/list-pairing.ts";
+import { type ListEntries } from "#src/tools/shared/validation/lists/list-pairing.ts";
 
 /** The names and colors a call hands out, and where the current source is. */
 export interface CopyLabels {
