+++ src/tools/shared/validation/color-utils.ts
@@ -8,7 +8,7 @@ import {
   splitList,
   valueForIndex,
   warnPairingMismatch,
-} from "#src/tools/shared/validation/list-pairing.ts";
+} from "#src/tools/shared/validation/lists/list-pairing.ts";
 
 /**
  * Parse a comma-separated color param and warn when it names the wrong number.
