+++ src/tools/shared/validation/lists/list-pairing.ts
