+++ src/tools/track/update/update-track.ts
@@ -31,11 +31,11 @@ import {
   getNameForIndex,
   parseNames,
 } from "#src/tools/shared/validation/name-utils.ts";
-import { validateListLengths } from "#src/tools/shared/validation/list-lengths.ts";
+import { validateListLengths } from "#src/tools/shared/validation/lists/list-lengths.ts";
 import {
   targetCount,
   targetIds,
-} from "#src/tools/shared/validation/target-lists.ts";
+} from "#src/tools/shared/validation/lists/target-lists.ts";
 import { trackIdPerPath } from "#src/tools/shared/validation/path-target-lookup.ts";
 
 interface UpdateTrackArgs {
